Import the test from its json format
This depends on:
- downloading the json file from the repo
- having jq installed
- the SML Unix structure working for you

We might soon easily avoid the last two by using a JSON parser in SML
instead.

Also, the test correctness is now cheated because this import revealed
the neglected beacon state root contract storage update in the post
state (the EVM semantics needs to be updated to include that).